// Pandem2-Prototype-API/src/interfaces/vaccines.ts
/*
  Copyright Clarisoft, a Modus Create Company, 20/07/2023, licensed under the
  EUPL-1.2 or later. This open-source code is licensed following the Attribution
  4.0 International (CC BY 4.0) - Creative Commons — Attribution 4.0 International
  — CC BY 4.0.

  Following this, you are accessible to:

  Share - copy and redistribute the material in any medium or format.
  Adapt - remix, transform, and build upon the material commercially.

  Remark: The licensor cannot revoke these freedoms if you follow the license
  terms.

  Under the following terms:

  Attribution - You must give appropriate credit, provide a link to the license,
  and indicate if changes were made. You may do so reasonably but not in any way
  that suggests the licensor endorses you or your use.
  No additional restrictions - You may not apply legal terms or technological
  measures that legally restrict others from doing anything the license permits.
*/
import { ILocation, TotalType } from './common';
import { DoseType, HealthcareWorker, VaccineGender } from '../models/vaccine';

export type VaccineSplitTypeQuery = 'dose_type' | 'gender' | 'age_group' | 'healthcare_worker' | 'population_type';

export interface IGenerateDummyDataPayload {
  location: ILocation;
  pathogenId: string;
  start_date: Date;
  end_date: Date;
  generate_for_sublocations?: boolean;
}

export interface IVaccinationsFilter {
  pathogenId?: string;
  location: string;
  split?: VaccineSplitTypeQuery;
  dose_type?: DoseType | DoseType[];
  total_type?: TotalType | TotalType[];
  gender?: VaccineGender;
  age_group?: string;
  healthcare_worker?: HealthcareWorker;
  population_type?: string;
  start_date?: string;
  end_date?: string;
}

export interface IDailyVaccinationsCount {
  date: string;
  total: number;
  split?: { total: number, split_value: string }[];
}

export interface ILocationsVaccinationsDateIntervalFilter {
  pathogenId?: string;
  location: string[];
  split?: VaccineSplitTypeQuery;
  dose_type?: DoseType;
  population_type?: string;
  gender?: VaccineGender;
  age_group?: string;
  healthcare_worker?: HealthcareWorker;
}